+++ app/models/weapon.rb
@@ -6,7 +6,7 @@ class Weapon < ApplicationRecord
   include CharacterTrait
 
   before_validation :trim_tags
-  before_validation :set_damage_for_elemental_bolt
+  before_validation :set_traits_for_elemental_bolt
 
   validates :weight, inclusion: { in: %w[ light medium heavy ] }
   validates :attr, inclusion: { in: Constants::ATTRIBUTES }
@@ -15,13 +15,16 @@ class Weapon < ApplicationRecord
   def trim_tags
     return unless will_save_change_to_attribute? :tags
 
-    self.tags = tags.reject(&:blank?).collect(&:strip).collect(&:downcase)
+    self.tags = tags.reject(&:blank?).collect(&:strip).collect(&:downcase).uniq
   end
 
-  def set_damage_for_elemental_bolt
+  def set_traits_for_elemental_bolt
     return unless will_save_change_to_attribute? :tags
 
-    self.damage_attr = 'essence' if tags.include? 'elemental bolt'
+    return unless tags.include?('elemental bolt') && !(tags_was.include? 'elemental bolt')
+
+    self.damage_attr = 'essence'
+    self.is_artifact = true
   end
 
   def entity_type