+++ app/serializers/weapon_serializer.rb
@@ -3,5 +3,6 @@
 # app/serializers/weapon_serializer.rb
 class WeaponSerializer < CharacterTraitSerializer
   attributes :name, :weight, :tags, :is_artifact,
-             :ability, :attr, :damage_attr
+             :ability, :attr, :damage_attr,
+             :bonus_accuracy, :bonus_damage, :bonus_defense, :bonus_overwhelming
 end