+++ db/migrate/20181001200806_add_bonuses_to_weapons.rb
@@ -1,0 +1,9 @@
+class AddBonusesToWeapons < ActiveRecord::Migration[5.2]
+  def change
+    add_column :weapons, :bonus_accuracy, :integer, default: 0
+    add_column :weapons, :bonus_damage, :integer, default: 0
+    add_column :weapons, :bonus_defense, :integer, default: 0
+    add_column :weapons, :bonus_overwhelming, :integer, default: 0
+    add_column :weapons, :notes, :text, default: ''
+  end
+end